[Refactor] refactoring for DPP
